// src/routes/mobile-observations/observer/types.ts
export type BasicAdData = {
	adId: string;
	timestamp: number;
	date: string;
	time: string;
	observer: string;
	path: string;
};

export type RichAdData = BasicAdData & {
	attributes?: {
		[key: string]: {
			value?: string;
			createdBy?: string;
			createdAt?: string;
			modifiedBy?: string;
			modifiedAt?: string;
		};
	};
};